nav logic refractor
redid frontend nav logic, sorted api route logic on back end, prepping
new pub profile section for SPA integration
