+++ routes.js
@@ -2,7 +2,9 @@ var helpers = require('./helpers');
 
 module.exports = (app) => {
   app.use('/',  require('./routes/index'));
-  app.use('/api',  helpers.middleware.checkxhr(), require('./routes/api'));
+  app.use('/api',  helpers.middleware.checkxhr(), require('./routes/api/index'));
+  app.use('/api',  helpers.middleware.checkxhr(), require('./routes/api/feedback'));
+  app.use('/api',  helpers.middleware.checkxhr(), require('./routes/api/general'));
   app.use('/faq',  require('./routes/faq'));
   app.use('/about',  require('./routes/about'));
   app.use('/disclaimer',  require('./routes/disclaimer'));
