+++ views/template.html
@@ -8,6 +8,8 @@
     <link rel="stylesheet" href="/css/style.css" media="screen" title="no title" charset="utf-8">
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/2.2.0/jquery.min.js"></script>
     <script src="/js/materialize.min.js" charset="utf-8"></script>
+    <script src="/js/signals.min.js" charset="utf-8"></script>
+    <script src="/js/crossroads.min.js" charset="utf-8"></script>
     <link href="http://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
   {% endblock %}
   {% block headaddons %}
